# backend/architecture/openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /user/preferences/extract:
    post:
      tags:
      - user
      summary: Preferences Extract
      operationId: preferences_extract_user_preferences_extract_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferenceExtractRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Preferences Extract User Preferences Extract Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /user/garments/refresh-description:
    post:
      tags:
      - user
      summary: Refresh Description
      operationId: refresh_description_user_garments_refresh_description_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshDescriptionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Refresh Description User Garments Refresh Description
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /garments/ingest:
    post:
      tags:
      - ingest
      summary: Garments Ingest
      description: 'Ingest a garment with an image (base64) and optional attributes.


        Returns garment_id and external_id. If the external_id was previously ingested,
        the

        existing id is returned (idempotent).'
      operationId: garments_ingest_garments_ingest_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Garments Ingest Garments Ingest Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IngestRequest:
      properties:
        external_id:
          type: string
          title: External Id
        image_base64:
          type: string
          title: Image Base64
        attributes:
          anyOf:
          - additionalProperties:
              items:
                type: string
              type: array
            type: object
          - type: 'null'
          title: Attributes
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        price:
          anyOf:
          - type: number
          - type: 'null'
          title: Price
        currency:
          anyOf:
          - type: string
          - type: 'null'
          title: Currency
      type: object
      required:
      - external_id
      - image_base64
      title: IngestRequest
...
    PreferenceExtractRequest:
      properties:
        conversation:
          type: string
          title: Conversation
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
      type: object
      required:
      - conversation
      title: PreferenceExtractRequest
    RefreshDescriptionRequest:
      properties:
        garment_id:
          type: integer
          title: Garment Id
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        overwrite:
          type: boolean
          title: Overwrite
          default: false
      type: object
      required:
      - garment_id
      title: RefreshDescriptionRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError